pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["himariseino <himawari1600ka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = {extras = ["voice"], version = "^2.4.0"}
python-dotenv = "^1.0.1"
langchain = "^0.3.9"
langchain-openai = "^0.2.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
